Calculate x/y min-max values for multi-selection.
